Write a C program that prints the size of various types on the computer it is compiled and run on.

+++ 0x00-hello_world/6-size.c
@@ -17,10 +17,10 @@ int main(void)
 
 	float f;
 
-	printf("size of char: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of int : %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of long: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("size of char: %lu byte(s)\n", sizeof(d));
+	printf("size of int : %lu byte(s)\n", sizeof(a));
+	printf("size of a long int: %lu byte(s)\n", sizeof(c));
+	printf("size of long: %lu byte(s)\n", sizeof(b));
+	printf("size of float: %lu byte(s)\n", sizeof(f));
 	return (0);
 }
